refactor: move translation string function to utils.js
